+++ src/com/projet/pharmatech/services/MedicamentService.java
@@ -17,7 +17,7 @@ public class MedicamentService {
 	}
 
 
-	public void delete(Long id) {
+	public void delete(int id) {
 		medicamentDao.delete(id);			
 	}
 
@@ -26,7 +26,7 @@ public class MedicamentService {
 		return medicamentDao.findAll();
 	}
 
-	public Medicament findById(Long id) {
+	public Medicament findById(int id) {
 		return medicamentDao.findById(id);
 	}
